+++ client/src/components/ActiveChat/Input.js
@@ -1,85 +1,67 @@
-import React, { Component } from "react";
-import { FormControl, FilledInput } from "@material-ui/core";
-import { withStyles } from "@material-ui/core/styles";
-import { connect } from "react-redux";
+import React, { useState } from "react";
+import { FormControl, FilledInput, ThemeProvider } from "@material-ui/core";
+import { useDispatch } from "react-redux";
 import { postMessage } from "../../store/utils/thunkCreators";
+import { makeStyles, createTheme } from "@material-ui/core/styles";
 
-const styles = {
+const useStyles = makeStyles(() => ({
   root: {
     justifySelf: "flex-end",
-    marginTop: 15,
   },
   input: {
     height: 70,
-    backgroundColor: "#F4F6FA",
     borderRadius: 8,
     marginBottom: 20,
-  },
-};
+},
+}));
 
-class Input extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      text: "",
-    };
-  }
+const theme = createTheme(
+  {
+    spacing:{
+      my: 1005,
+    },
+    palette: {
+      main: "#F4F6FA",
+    }
+  });
 
-  handleChange = (event) => {
-    this.setState({
-      text: event.target.value,
-    });
+const Input =(props) => {
+  const [text, setText] = useState('')
+  const classes = useStyles();
+  const dispatch = useDispatch();
+  const handleChange = (event) => {
+    setText(event.target.value);
   };
 
-  handleSubmit = async (event) => {
+  const handleSubmit = async (event) => {
     event.preventDefault();
     // add sender user info if posting to a brand new convo, so that the other user will have access to username, profile pic, etc.
     const reqBody = {
-      text: event.target.text.value,
-      recipientId: this.props.otherUser.id,
-      sender: this.props.conversationId ? null : this.props.user,
+      text,
+      recipientId: props.otherUser.id,
+      sender: props.user,
     };
-    await this.props.postMessage(reqBody);
-    this.setState({
-      text: "",
-    });
+    await dispatch(postMessage(reqBody));
+    setText('')
   };
 
-  render() {
-    const { classes } = this.props;
+
     return (
-      <form className={classes.root} onSubmit={this.handleSubmit}>
-        <FormControl fullWidth hiddenLabel>
-          <FilledInput
-            classes={{ root: classes.input }}
-            disableUnderline
-            placeholder="Type something..."
-            value={this.state.text}
-            name="text"
-            onChange={this.handleChange}
-          />
-        </FormControl>
-      </form>
+      <ThemeProvider theme={theme}>
+        <form className={classes.root} onSubmit={handleSubmit}>
+          <FormControl fullWidth hiddenLabel>
+            <FilledInput
+              classes={{ root: classes.input }}
+              disableUnderline
+              placeholder="Type something..."
+              value={text}
+              name="text"
+              onChange={handleChange}
+            />
+          </FormControl>
+        </form>
+      </ThemeProvider>
     );
   }
-}
 
-const mapStateToProps = (state) => {
-  return {
-    user: state.user,
-    conversations: state.conversations,
-  };
-};
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    postMessage: (message) => {
-      dispatch(postMessage(message));
-    },
-  };
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(withStyles(styles)(Input));
+export default Input;
