+++ client/src/components/ActiveChat/index.js
@@ -4,4 +4,4 @@ export { default as ActiveChat } from "./ActiveChat";
 export { default as Input } from "./Input";
 export { default as Header } from "./Header";
 export { default as Messages } from "./Messages";
-export { default as FunctionalInput } from "./FunctionalInput"
+
