+++ client/src/components/Sidebar/Search.js
@@ -1,16 +1,12 @@
-import React, { Component } from "react";
-import { FormControl, FilledInput, InputAdornment } from "@material-ui/core";
+import React from "react";
+import { FormControl, FilledInput, InputAdornment, ThemeProvider } from "@material-ui/core";
 import SearchIcon from "@material-ui/icons/Search";
-import { withStyles } from "@material-ui/core/styles";
+import { makeStyles, createTheme } from "@material-ui/core/styles";
 
-const styles = {
+const useStyles = makeStyles(() => ({
   filledInput: {
     height: 50,
     background: "#E9EEF9",
-    borderRadius: 5,
-    fontSize: 13,
-    fontWeight: "bold",
-    color: "#99A9C4",
     letterSpacing: 0,
     display: "flex",
     justifyContent: "center",
@@ -22,34 +18,41 @@ const styles = {
       opacity: 1,
     },
   },
-};
+}));
+const theme = createTheme(
+  {
+    typography:{
+      fontSize: 13,
+      fontWeight: "bold",
+    }
+  });
 
-class Search extends Component {
-  handleSubmit = (event) => {
+const Search = (props) =>{
+  const handleSubmit = (event) => {
     event.preventDefault();
   };
+  const classes = useStyles()
 
-  render() {
-    const { classes } = this.props;
     return (
-      <form onSubmit={this.handleSubmit}>
-        <FormControl fullWidth hiddenLabel>
-          <FilledInput
-            name="search"
-            onChange={this.props.handleChange}
-            classes={{ root: classes.filledInput, input: classes.input }}
-            disableUnderline
-            placeholder="Search"
-            startAdornment={
-              <InputAdornment position="start">
-                <SearchIcon />
-              </InputAdornment>
-            }
-          ></FilledInput>
-        </FormControl>
+      <form onSubmit={handleSubmit}>
+        <ThemeProvider theme={theme}>
+          <FormControl fullWidth hiddenLabel>
+              <FilledInput
+                name="search"
+                onChange={props.handleChange}
+                classes={{ root: classes.filledInput, input: classes.input }}
+                disableUnderline
+                placeholder="Search"
+                startAdornment={
+                  <InputAdornment position="start">
+                    <SearchIcon />
+                  </InputAdornment>
+               }
+              ></FilledInput>
+            </FormControl>
+          </ThemeProvider>
       </form>
     );
-  }
 }
 
-export default withStyles(styles)(Search);
+export default Search;
