+++ client/src/components/Sidebar/index.js
@@ -5,5 +5,5 @@ export { default as Sidebar } from "./Sidebar";
 export { default as SidebarContainer } from "./SidebarContainer";
 export { default as Chat } from "./Chat";
 export { default as BadgeAvatar } from "./BadgeAvatar";
-export { default as FunctionalSearch } from "./FunctionalSearch"
-export { default as FunctionalChat } from "./FunctionalChat"
+
+
